feat: split up prover bin
