+++ Cargo.toml
@@ -10,7 +10,7 @@ repository = "https://github.com/worldcoin/world-chain/"
 resolver = "3"
 members = [
     "bin/world-chain",
-    "proofs/bin",
+    "proofs/prover",
     "crates/chainspec",
     "crates/node",
     "crates/rpc",
@@ -28,6 +28,8 @@ members = [
     "proofs/protocol",
     "proofs/nitro",
     "proofs/nitro/worker",
+    "proofs/prover-sp1",
+    "proofs/prover-nitro",
     "proofs/kona-client",
     "proofs/kona-host",
     "proofs/succinct/utils/client",
@@ -87,6 +89,9 @@ world-chain-evm = { path = "crates/evm", default-features = false }
 world-chain-state = { path = "crates/state", default-features = false }
 world-chain-devnet = { path = "crates/devnet", default-features = false }
 world-chain-test-utils = { path = "crates/test-utils", default-features = false }
+world-chain-prover = { path = "proofs/prover", default-features = false }
+world-chain-prover-sp1 = { path = "proofs/prover-sp1", default-features = false }
+world-chain-prover-nitro = { path = "proofs/prover-nitro", default-features = false }
 world-chain-proof-core = { path = "proofs/core", default-features = false }
 world-chain-proof-protocol = { path = "proofs/protocol", default-features = false }
 world-chain-proof-nitro = { path = "proofs/nitro", default-features = false }
