+++ Dockerfile.prover
@@ -1,7 +1,7 @@
 # syntax=docker/dockerfile:1.10
-# Image for the host-side prover binaries (proofs/*). Defaults to the `proof` CLI built
-# with the sp1 + nitro backends; pass PROVER_PACKAGE/PROVER_BIN/FEATURES to build others
-# (e.g. PROVER_PACKAGE=world-chain-sp1-worker PROVER_BIN=sp1-worker FEATURES="").
+# Image for host-side prover binaries. Defaults to the SP1 prover; pass
+# PROVER_BACKEND=nitro to build the Nitro prover. Service images can still pass
+# PROVER_PACKAGE/PROVER_BIN directly.
 # Mirrors the caching strategy of the main Dockerfile (cargo-chef + sccache/S3).
 FROM public.ecr.aws/docker/library/rust:1.95.0-bookworm AS base
 
@@ -34,10 +34,10 @@ RUN --mount=type=cache,target=/usr/local/cargo/registry \
 FROM base AS builder
 WORKDIR /app
 
-ARG PROVER_PACKAGE="proof"
-ARG PROVER_BIN="proof"
+ARG PROVER_BACKEND="sp1"
+ARG PROVER_PACKAGE=""
+ARG PROVER_BIN=""
 ARG PROFILE="maxperf"
-ARG FEATURES="sp1,nitro"
 ARG SCCACHE_BUCKET
 ARG SCCACHE_REGION
 ARG SCCACHE_S3_KEY_PREFIX
@@ -51,7 +51,14 @@ RUN --mount=type=cache,target=/usr/local/cargo/registry \
     --mount=type=secret,id=aws_secret_access_key,env=AWS_SECRET_ACCESS_KEY \
     --mount=type=secret,id=aws_session_token,env=AWS_SESSION_TOKEN \
     if [ -z "$SCCACHE_BUCKET" ]; then unset SCCACHE_BUCKET SCCACHE_REGION SCCACHE_S3_KEY_PREFIX; fi && \
-    cargo chef cook --locked --profile ${PROFILE} -p ${PROVER_PACKAGE} --bin ${PROVER_BIN} ${FEATURES:+--features ${FEATURES}} --recipe-path recipe.json
+    case "$PROVER_BACKEND" in \
+      sp1) DEFAULT_PROVER_PACKAGE="world-chain-prover-sp1"; DEFAULT_PROVER_BIN="world-chain-prover-sp1" ;; \
+      nitro) DEFAULT_PROVER_PACKAGE="world-chain-prover-nitro"; DEFAULT_PROVER_BIN="world-chain-prover-nitro" ;; \
+      *) echo "unsupported PROVER_BACKEND: $PROVER_BACKEND" >&2; exit 1 ;; \
+    esac && \
+    PROVER_PACKAGE="${PROVER_PACKAGE:-$DEFAULT_PROVER_PACKAGE}" && \
+    PROVER_BIN="${PROVER_BIN:-$DEFAULT_PROVER_BIN}" && \
+    cargo chef cook --locked --profile ${PROFILE} -p ${PROVER_PACKAGE} --bin ${PROVER_BIN} --recipe-path recipe.json
 
 ARG VERGEN_GIT_SHA
 COPY . .
@@ -63,7 +70,16 @@ RUN --mount=type=cache,target=/usr/local/cargo/registry \
     --mount=type=secret,id=aws_secret_access_key,env=AWS_SECRET_ACCESS_KEY \
     --mount=type=secret,id=aws_session_token,env=AWS_SESSION_TOKEN \
     if [ -z "$SCCACHE_BUCKET" ]; then unset SCCACHE_BUCKET SCCACHE_REGION SCCACHE_S3_KEY_PREFIX; fi && \
-    cargo build --locked --profile ${PROFILE} -p ${PROVER_PACKAGE} ${FEATURES:+--features ${FEATURES}} --bin ${PROVER_BIN}
+    case "$PROVER_BACKEND" in \
+      sp1) DEFAULT_PROVER_PACKAGE="world-chain-prover-sp1"; DEFAULT_PROVER_BIN="world-chain-prover-sp1" ;; \
+      nitro) DEFAULT_PROVER_PACKAGE="world-chain-prover-nitro"; DEFAULT_PROVER_BIN="world-chain-prover-nitro" ;; \
+      *) echo "unsupported PROVER_BACKEND: $PROVER_BACKEND" >&2; exit 1 ;; \
+    esac && \
+    PROVER_PACKAGE="${PROVER_PACKAGE:-$DEFAULT_PROVER_PACKAGE}" && \
+    PROVER_BIN="${PROVER_BIN:-$DEFAULT_PROVER_BIN}" && \
+    cargo build --locked --profile ${PROFILE} -p ${PROVER_PACKAGE} --bin ${PROVER_BIN} && \
+    cp "target/${PROFILE}/${PROVER_BIN}" /tmp/prover-bin && \
+    printf '%s\n' "$PROVER_BIN" > /tmp/prover-bin-name
 
 FROM public.ecr.aws/docker/library/debian:bookworm-slim
 WORKDIR /app
@@ -75,10 +91,12 @@ RUN apt-get update && \
   apt-get clean && \
   rm -rf /var/lib/apt/lists/*
 
-ARG PROVER_BIN="proof"
-ARG PROFILE="maxperf"
-COPY --from=builder /app/target/${PROFILE}/${PROVER_BIN} /usr/local/bin/${PROVER_BIN}
-RUN ln -s /usr/local/bin/${PROVER_BIN} /usr/local/bin/entrypoint
+COPY --from=builder /tmp/prover-bin /usr/local/bin/entrypoint
+COPY --from=builder /tmp/prover-bin-name /tmp/prover-bin-name
+RUN chmod +x /usr/local/bin/entrypoint && \
+  bin_name="$(cat /tmp/prover-bin-name)" && \
+  ln -s /usr/local/bin/entrypoint "/usr/local/bin/${bin_name}" && \
+  rm /tmp/prover-bin-name
 
 # Committed SP1 guest ELFs, so prover binaries that load ELFs at runtime work
 # out of the box. The env vars are the canonical lookup used by the hosts.
