+++ Justfile
@@ -99,7 +99,7 @@ build-proof-elfs: build-proof-range-elf build-proof-aggregation-elf
 
 # Compute the on-chain verification keys for the committed SP1 proof ELFs
 proof-vkeys *args='':
-    cargo run --release -p proof --features sp1 -- sp1 vkeys $@
+    cargo run --release -p world-chain-prover-sp1 -- vkeys $@
 
 # Generate CLI reference docs for the mdbook
 docs:
