--- proofs/bin/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "proof"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[[bin]]
-name = "proof"
-path = "src/main.rs"
-
-[dependencies]
-alloy-primitives.workspace = true
-anyhow.workspace = true
-clap.workspace = true
-dotenvy.workspace = true
-hex.workspace = true
-reqwest = { workspace = true, features = ["blocking", "json", "rustls-tls"] }
-rkyv.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-world-chain-chainspec.workspace = true
-world-chain-proof-core.workspace = true
-world-chain-proof-protocol.workspace = true
-world-chain-proof-kona-host-utils.workspace = true
-world-chain-proof-succinct-host-utils.workspace = true
-sp1-sdk = { version = "=6.1.0", features = ["network"], optional = true }
-sha2 = { workspace = true, optional = true }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-world-chain-proof-nitro = { workspace = true, optional = true }
-
-[features]
-default = []
-nitro = ["dep:world-chain-proof-nitro", "world-chain-proof-nitro/aws_nitro"]
-sp1 = ["dep:sp1-sdk", "dep:sha2", "world-chain-proof-succinct-host-utils/sp1"]
