+++ proofs/prover-nitro/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "world-chain-prover-nitro"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[[bin]]
+name = "world-chain-prover-nitro"
+path = "src/main.rs"
+
+[dependencies]
+anyhow.workspace = true
+clap.workspace = true
+dotenvy.workspace = true
+hex.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+world-chain-prover.workspace = true
+
+[target.'cfg(target_os = "linux")'.dependencies]
+world-chain-proof-nitro = { workspace = true, features = ["aws_nitro"] }
