+++ proofs/prover-sp1/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "world-chain-prover-sp1"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[[bin]]
+name = "world-chain-prover-sp1"
+path = "src/main.rs"
+
+[dependencies]
+alloy-primitives.workspace = true
+anyhow.workspace = true
+clap.workspace = true
+dotenvy.workspace = true
+hex.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
+sp1-sdk = { version = "=6.1.0", features = ["network"] }
+tokio.workspace = true
+world-chain-prover.workspace = true
+world-chain-proof-core.workspace = true
+world-chain-proof-succinct-host-utils = { workspace = true, features = ["sp1"] }
