+++ proofs/prover/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "world-chain-prover"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+alloy-primitives.workspace = true
+anyhow.workspace = true
+clap.workspace = true
+reqwest = { workspace = true, features = ["blocking", "json", "rustls-tls"] }
+rkyv.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+world-chain-chainspec.workspace = true
+world-chain-proof-core.workspace = true
+world-chain-proof-protocol.workspace = true
+world-chain-proof-kona-host-utils.workspace = true
